Adicionada a validação de itens vinculados ao pedido, para evitar remoção de pedidos com itens vinculados.
Changes to be committed:
	modified:   src/controlles/PedidoController.js
	modified:   src/models/modelContador.js
	modified:   src/models/modelPedido.js
	modified:   src/services/ItensPedidoService.js

+++ src/controlles/PedidoController.js
@@ -26,8 +26,14 @@ module.exports = class PedidoController {
 			return ;
 		}
 
-		const pedido = await PedidoService.deleteOne(req.params.numPed);
-		res.json(pedido);
+		const exists = await ItensPedidoService.findOne(req.params.numPed);
+		console.log(exists);
+		if (!exists) {
+			const pedido = await PedidoService.deleteOne(req.params.numPed);
+			res.json(pedido);
+			return ;
+		}
+		res.status(400).json({ erro:'erro ao deletar pedido, ha itens vinculados' })
 	}
 
 	async insertProductPed(req, res) {
